tier1/pyproject.toml
[project]
name = "tier1"
version = "0.1.0"
description = ""
authors = [
    {name = "Samuel Stidham",email = "samuel.stidham@snhu.edu"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "twilio (>=9.8.4,<10.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "tier1", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "pytest (>=8.4.2,<9.0.0)"
]
